requirements.txt
ImageHash==4.3.1
loguru==0.7.2
m3u8==4.0.0
memory-profiler==0.61.0
pillow>=10.2.0
psutil==5.9.8
pyffmpeg>=2.5.2.3.1
python-dateutil==2.8.2
requests==2.31.0
rich==13.7.0
strenum==0.4.15
websockets==12.0

# GUI Dependencies
customtkinter>=5.2.0
darkdetect>=0.8.0
tkinterdnd2>=0.3.0

# Image processing dependencies
numpy>=1.24.0

# Enhanced compression features
scikit-image>=0.21.0  # SSIM quality validation
mozjpeg-lossless-optimization>=1.1.2  # MozJPEG lossless optimization
